+++ library/src/repositories/ServiceRepository.cpp
@@ -206,7 +206,7 @@ namespace RehabClinic
 						throw EmptyRecordException("Equipment");
 					};
 
-					for (i = 0; eqp.size() - 1 > i; i++)
+					for (i = 0; (int)eqp.size() - 1 > i; i++)
 					{
 						outFile << static_cast<int>(rehabilitationService->getRequiredEqupiment()[i]) << ",";
 					}
